+++ core/src/main/kotlin/net/corda/core/node/PluginServiceHub.kt
@@ -1,6 +1,0 @@
-package net.corda.core.node
-
-/**
- * A service hub to be used by the [CordaPluginRegistry]
- */
-interface PluginServiceHub : ServiceHub