+++ core/src/main/kotlin/net/corda/core/node/services/CordaService.kt
@@ -1,6 +1,5 @@
 package net.corda.core.node.services
 
-import net.corda.core.node.PluginServiceHub
 import net.corda.core.node.ServiceHub
 import net.corda.core.serialization.SerializeAsToken
 import net.corda.core.serialization.SingletonSerializeAsToken
@@ -8,9 +7,9 @@ import kotlin.annotation.AnnotationTarget.CLASS
 
 /**
  * Annotate any class that needs to be a long-lived service within the node, such as an oracle, with this annotation.
- * Such a class needs to have a constructor with a single parameter of type [PluginServiceHub]. This construtor will be
- * invoked during node start to initialise the service. The service hub provided can be used to get information about the
- * node that may be necessary for the service. Corda services are created as singletons within the node and are available
+ * Such a class needs to have a constructor with a single parameter of type [ServiceHub]. This construtor will be invoked
+ * during node start to initialise the service. The service hub provided can be used to get information about the node
+ * that may be necessary for the service. Corda services are created as singletons within the node and are available
  * to flows via [ServiceHub.cordaService].
  *
  * The service class has to implement [SerializeAsToken] to ensure correct usage within flows. (If possible extend